cloudSnip/processing/pyproject.toml
[project]
name = "cloudsnip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dask>=2025.5.1",
    "distributed>=2025.5.1",
    "ephem>=4.2",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "notebook>=7.4.3",
    "pillow>=11.2.1",
    "rioxarray>=0.19.0",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
